backend/requirements.txt
alembic==1.13.1                             
annotated-types==0.7.0                      
anyio==3.7.1                                
bcrypt==5.0.0                               
certifi==2026.2.25
cffi>=1.17.1
charset-normalizer==3.4.7
click==8.1.8                                
colorama==0.4.6                             
cryptography==46.0.6
ecdsa==0.19.2
elevenlabs>=1.0.0,<2.0.0
exceptiongroup==1.3.1
fastapi==0.104.1
gunicorn==21.2.0
google-ai-generativelanguage==0.6.0
google-api-core==2.29.0
google-auth==2.49.1
google-generativeai==0.3.2
googleapis-common-protos==1.73.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Mako==1.3.10
MarkupSafe==2.1.5
msgpack==1.1.1
passlib==1.7.4
proto-plus==1.27.1
protobuf==4.25.9
py-algorand-sdk==2.4.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.4.2
pydantic-settings>=2.1.0
pydantic-core==2.10.1
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytimeparse==1.1.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.24.0
websockets==13.1
zipp==3.20.2
psycopg2-binary==2.9.9